UseReplicationScope state and better organisation
