+++ replication/src/rooms/params.rs
@@ -1,28 +1,28 @@
 use bevy::{ecs::{query::QueryEntityError, system::SystemParam}, prelude::*};
 use bevy_stardust::prelude::*;
 use crate::prelude::*;
-use super::RoomsEnabled;
+use super::UseReplicationScope;
 
 /// Utilities relating to replication scope, such as determining if an entity is in scope to a peer.
 #[derive(SystemParam)]
 pub struct ReplicationScope<'w, 's> {
-    enabled: Option<Res<'w, RoomsEnabled>>,
+    enabled: Res<'w, State<UseReplicationScope>>,
     peers: Query<'w, 's, Entity, (With<NetworkPeer>, With<ReplicationPeer>)>,
-    groups: Query<'w, 's, (&'static NetworkRoom, &'static NetworkGroup)>,
-    entities: Query<'w, 's, &'static ReplicateEntity>,
+    groups: Query<'w, 's, (Entity, &'static NetworkRoom, &'static NetworkGroup)>,
+    entities: Query<'w, 's, (Entity, &'static ReplicateEntity)>,
 }
 
 impl<'w, 's> ReplicationScope<'w, 's> {
     /// Returns `true` if rooms and replication scoping is enabled.
     #[inline]
     pub fn rooms_enabled(&self) -> bool {
-        self.enabled.is_some()
+        self.enabled.get().is_enabled()
     }
 
     /// Returns `true` if the peer is in `room`.
     pub fn peer_in_room(&self, peer: Entity, room: Entity) -> Result<bool, QueryEntityError> {
         if self.peers.get(peer).is_err() { return Err(QueryEntityError::NoSuchEntity(peer)); }
-        let (_, group) = self.groups.get(room)?;
+        let (_, _, group) = self.groups.get(room)?;
         return Ok(group.contains(peer));
     }
